backend/requirements.txt
uvicorn>=0.27.0,<0.28.0
fastapi>=0.110.0,<0.111.0
sqlalchemy>=2.0.0,<2.1.0
passlib>=1.7.0,<1.8.0
python-jose>=3.3.0,<3.4.0
